Defer creating MethodInvocation info that uses the VirtualExecutorUtil to avoid loading problems if JDK does not include ThreadPerTaskExecutor
