+++ common/configurable/src/main/java/io/helidon/common/configurable/ObserverManager.java
@@ -15,6 +15,7 @@
  */
 package io.helidon.common.configurable;
 
+import java.lang.reflect.Method;
 import java.util.List;
 import java.util.Map;
 import java.util.ServiceLoader;
@@ -51,13 +52,20 @@ class ObserverManager {
 
     private static final Logger LOGGER = Logger.getLogger(ObserverManager.class.getName());
 
-    private static final LazyValue<List<ExecutorServiceSupplierObserver>> OBSERVER_LIST = LazyValue
+    private static final LazyValue<List<ExecutorServiceSupplierObserver>> OBSERVERS = LazyValue
             .create(ObserverManager::loadObservers);
 
     private static final Map<Supplier<? extends ExecutorService>, SupplierInfo> SUPPLIERS = new ConcurrentHashMap<>();
-    private static final Map<String, AtomicInteger> SUPPLIER_NEXT_INDEX_VALUES = new ConcurrentHashMap<>();
+
+    // A given supplier category can have multiple suppliers, so keep track of the next available index by category.
+    private static final Map<String, AtomicInteger> SUPPLIER_CATEGORY_NEXT_INDEX_VALUES = new ConcurrentHashMap<>();
 
     private static final Map<ExecutorService, SupplierInfo> EXECUTOR_SERVICES = new ConcurrentHashMap<>();
+
+    // Defer building list until use to avoid loading problems if this JDK does not support ThreadPerTasExecutor.
+    private static final LazyValue<List<ExecutorServiceSupplierObserver.MethodInvocation>> METRICS_RELATED_METHOD_INVOCATIONS =
+            LazyValue.create(() -> List.of(
+                    MethodInvocationImpl.create("Thread count", "thread-count", "threadCount")));
 
     private ObserverManager() {
     }
@@ -74,7 +82,7 @@ class ObserverManager {
                                  String supplierCategory,
                                  String executorServiceCategory,
                                  boolean useVirtualThreads) {
-        int supplierIndex = SUPPLIER_NEXT_INDEX_VALUES.computeIfAbsent(supplierCategory, key -> new AtomicInteger())
+        int supplierIndex = SUPPLIER_CATEGORY_NEXT_INDEX_VALUES.computeIfAbsent(supplierCategory, key -> new AtomicInteger())
                 .getAndIncrement();
         SUPPLIERS.computeIfAbsent(supplier,
                                   s -> SupplierInfo.create(s,
@@ -97,6 +105,15 @@ class ObserverManager {
         registerSupplier(supplier, supplierCategory, executorServiceCategory, false);
     }
 
+    /**
+     * Registers an executor service from a supplier.
+     *
+     * @param supplier the supplier registering the executor service
+     * @param executorService the executor service being registered
+     * @param <E> type of the executor service being registered
+     * @return the same executor service being registered
+     * @throws IllegalStateException if the supplier has not previously registered itself
+     */
     static <E extends ExecutorService> E registerExecutorService(Supplier<E> supplier, E executorService) {
         SupplierInfo supplierInfo = SUPPLIERS.get(supplier);
         if (supplierInfo == null) {
@@ -106,7 +123,15 @@ class ObserverManager {
         return executorService;
     }
 
-    static <E extends ExecutorService> void unregisterExecutorService(E executorService) {
+    /**
+     * Unregisters a previously-registered executor service that is being shut down.
+     * <p>
+     *     During production, the executor service would have been previously registered by a supplier. But during testing that
+     *     is not always the case.
+     * </p>
+     * @param executorService the executor service being shut down
+     */
+    static void unregisterExecutorService(ExecutorService executorService) {
         SupplierInfo supplierInfo = EXECUTOR_SERVICES.get(executorService);
         if (supplierInfo == null) {
             // This can happen in some unit tests but should not happen in production.
@@ -156,14 +181,14 @@ class ObserverManager {
         }
 
         private List<ExecutorServiceSupplierObserver.SupplierObserverContext> collectObserverContexts() {
-            return OBSERVER_LIST.get()
+            return OBSERVERS.get()
                     .stream()
                     .map(observer ->
                                  useVirtualThreads
                                          ? observer.registerSupplier(supplier,
                                                                      supplierIndex,
                                                                      supplierCategory,
-                                                                     VirtualExecutorUtil.METRICS_RELATED_METHOD_INVOCATIONS)
+                                                                     METRICS_RELATED_METHOD_INVOCATIONS.get())
                                          : observer.registerSupplier(supplier,
                                                                      supplierIndex,
                                                                      supplierCategory))
@@ -182,6 +207,54 @@ class ObserverManager {
             observerContexts
                     .forEach(observer -> observer.unregisterExecutorService(executorService));
             EXECUTOR_SERVICES.remove(executorService);
+        }
+    }
+
+    private static class MethodInvocationImpl implements ExecutorServiceSupplierObserver.MethodInvocation {
+        private final String displayName;
+        private final String description;
+        private final Method method;
+        private final Class<?> type;
+
+        private static final LazyValue<ExecutorService> VIRTUAL_EXECUTOR_SERVICE = LazyValue
+                .create(VirtualExecutorUtil::executorService);
+
+        static MethodInvocationImpl create(String displayName, String description, String methodName)  {
+            ExecutorService executorService = VIRTUAL_EXECUTOR_SERVICE.get();
+            Method method = null;
+            try {
+                method = executorService.getClass().getDeclaredMethod(methodName);
+            } catch (NoSuchMethodException e) {
+                throw new RuntimeException(e);
+            }
+            return new MethodInvocationImpl(displayName, description, method);
+        }
+
+        MethodInvocationImpl(String displayName, String description, Method method) {
+            this.displayName = displayName;
+            this.description = description;
+            this.method = method;
+            this.type = method.getReturnType();
+        }
+
+        @Override
+        public String displayName() {
+            return displayName;
+        }
+
+        @Override
+        public String description() {
+            return description;
+        }
+
+        @Override
+        public Method method() {
+            return method;
+        }
+
+        @Override
+        public Class<?> type() {
+            return type;
         }
     }
 }
